+++ packages/ui5-config/src/ui5-config.ts
@@ -1,5 +1,11 @@
-import { MiddlewareConfig } from './types';
-import { YamlDocument, NodeComment } from '@sap-ux/yaml';
+import { FioriToolsProxyConfig, ProxyBackend, ProxyUIConfig } from './types';
+import { YamlDocument, NodeComment, YAMLMap } from '@sap-ux/yaml';
+import {
+    getAppReloadMiddlewareConfig,
+    getFioriToolsProxyMiddlewareConfig,
+    getMockServerMiddlewareConfig
+} from './middlewares';
+import { CustomMiddleware } from 'index';
 
 /**
  * Represents a UI5 config file in yaml format (ui5(-*).yaml) with utility functions to manipulate the yaml document.
@@ -26,24 +32,29 @@ export class UI5Config {
     /**
      * Adds a UI5 Framework entry to the yaml file.
      *
+     * @param {string} ui5Framework - whether to user SAPUI5 or OpenUI5
      * @param {string} ui5Version - ui5 version
-     * @param {string[]} libraries - a list of libraries
-     * @param ui5Theme - ui5 theme
+     * @param {string[]} ui5Libraries - a list of libraries
+     * @param {string} ui5Theme - optional ui5 theme
      * @returns {UI5Config} the UI5Config instance
      * @memberof UI5Config
      */
-    public addUI5Framework(ui5Version: string, libraries: string[], ui5Theme = 'sap_fiori_3'): UI5Config {
+    public addUI5Framework(
+        ui5Framework: string,
+        ui5Version: string,
+        ui5Libraries: string[],
+        ui5Theme = 'sap_fiori_3'
+    ): UI5Config {
         const libraryObjs = [];
-        for (const library of libraries) {
+        for (const library of ui5Libraries) {
             libraryObjs.push({ name: library });
         }
-        // Add theme lib
-        ui5Theme = ui5Theme.replace(/_dark$/, ''); // Dark theme versions are provided by base theme lib
-        libraryObjs.push({ name: `themelib_${ui5Theme}` });
+        // Add theme lib (dark theme versions are provided by base theme lib)
+        libraryObjs.push({ name: `themelib_${ui5Theme.replace(/_dark$/, '')}` });
 
         this.document.setIn({
             path: 'framework',
-            value: { name: 'SAPUI5', version: ui5Version, libraries: libraryObjs }
+            value: { name: ui5Framework, version: ui5Version, libraries: libraryObjs }
         });
         return this;
     }
@@ -51,15 +62,118 @@ export class UI5Config {
     /**
      * Adds a list of custom middlewares to the config.
      *
-     * @param {MiddlewareConfig[]} middlewares - the list of custom middlewares
+     * @param {CustomMiddleware<any>[]} middlewares - the list of custom middlewares
      * @param {NodeComment<MiddlewareConfig>[]} [comments] - a list of comments
      * @returns {UI5Config} the UI5Config instance
      * @memberof UI5Config
      */
-    public addCustomMiddleware(middlewares: MiddlewareConfig[], comments?: NodeComment<MiddlewareConfig>[]): UI5Config {
+    public addCustomMiddleware(
+        middlewares: CustomMiddleware<any>[],
+        comments?: NodeComment<CustomMiddleware<any>>[]
+    ): UI5Config {
         for (const mw of middlewares) {
             this.document.appendTo({ path: 'server.customMiddleware', value: mw, comments });
         }
+        return this;
+    }
+
+    /**
+     * Adds a instance of the Fiori tools app-reload middleware to the config.
+     *
+     * @returns {UI5Config} the UI5Config instance
+     * @memberof UI5Config
+     */
+    public addFioriToolsAppReloadMiddleware(): UI5Config {
+        this.document.appendTo({
+            path: 'server.customMiddleware',
+            value: getAppReloadMiddlewareConfig()
+        });
+        return this;
+    }
+
+    /**
+     * Adds a instance of the Fiori tools proxy middleware to the config.
+     *
+     * @param proxyConfig proxy configuration containing an optional array of backend and an option UI5 host configuration
+     * @returns {UI5Config} the UI5Config instance
+     * @memberof UI5Config
+     */
+    public addFioriToolsProxydMiddleware(proxyConfig: FioriToolsProxyConfig): UI5Config {
+        const { config, comments } = getFioriToolsProxyMiddlewareConfig(proxyConfig.backend, proxyConfig.ui5);
+        this.document.appendTo({
+            path: 'server.customMiddleware',
+            value: config,
+            comments: comments as any
+        });
+        return this;
+    }
+
+    /**
+     * Adds a backend configuration to an existing fiori-tools-proxy middleware. If the config does not contain a fiori-tools-proxy middleware, an error is thrown.
+     *
+     * @param backend config of backend that is to be proxied
+     * @returns {UI5Config} the UI5Config instance
+     * @memberof UI5Config
+     */
+    public addBackendToFioriToolsProxydMiddleware(backend: ProxyBackend): UI5Config {
+        const middlewareList = this.document.getSequence({ path: 'server.customMiddleware' });
+        const proxyMiddleware = this.document.findItem(
+            middlewareList,
+            (item: any) => item.name === 'fiori-tools-proxy'
+        );
+        if (!proxyMiddleware) {
+            throw new Error('Could not find fiori-tools-proxy');
+        }
+        this.document.getMap({ start: proxyMiddleware as YAMLMap, path: 'configuration' }).set('backend', [backend]);
+        return this;
+    }
+
+    /**
+     * Adds a ui configuration to an existing fiori-tools-proxy middleware. If the config does not contain a fiori-tools-proxy middleware, an error is thrown.
+     *
+     * @param ui5 config of backend that is to be proxied
+     * @returns {UI5Config} the UI5Config instance
+     * @memberof UI5Config
+     */
+    public addUi5ToFioriToolsProxydMiddleware(ui5: ProxyUIConfig): UI5Config {
+        try {
+            const middlewareList = this.document.getSequence({ path: 'server.customMiddleware' });
+            const proxyMiddleware = this.document.findItem(
+                middlewareList,
+                (item: any) => item.name === 'fiori-tools-proxy'
+            );
+            if (proxyMiddleware && ui5 !== undefined) {
+                const configurationUi5Doc = this.document.getMap({
+                    start: proxyMiddleware as YAMLMap,
+                    path: 'configuration.ui5'
+                });
+                if (ui5.version) {
+                    configurationUi5Doc.set('version', ui5.version);
+                }
+                if (ui5.url) {
+                    configurationUi5Doc.set('url', ui5.url);
+                }
+                if (ui5.directLoad) {
+                    configurationUi5Doc.set('directLoad', ui5.directLoad);
+                }
+            }
+        } catch (e) {
+            // Ignore
+        }
+        return this;
+    }
+    /**
+     * Adds a instance of the mockserver middleware to the config.
+     *
+     * @param path option path that is to be mocked
+     * @returns {UI5Config} the UI5Config instance
+     * @memberof UI5Config
+     */
+    public addMockServerMiddleware(path?: string): UI5Config {
+        this.document.appendTo({
+            path: 'server.customMiddleware',
+            value: getMockServerMiddlewareConfig(path)
+        });
         return this;
     }