+++ packages/fiori-freestyle-writer/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @sap-ux/fiori-freestyle-writer
 
+## 0.10.2
+
+### Patch Changes
+
+-   3d7b9f6: Use exact versions of dependencies
+
+## 0.10.1
+
+### Patch Changes
+
+-   a7670a0: Update ui5 verison handling for ui5.yaml. Allow setting view name in basic template, enhancements to yaml lib.
+-   Updated dependencies [a7670a0]
+    -   @sap-ux/odata-service-writer@0.9.5
+    -   @sap-ux/ui5-application-writer@0.9.5
+    -   @sap-ux/ui5-config@0.10.1
+
+## 0.10.0
+
+### Minor Changes
+
+-   d37c8bd: Added support for selecting a custom view name for Fiori freestyle - Basic template
+
+### Patch Changes
+
+-   Updated dependencies [d37c8bd]
+    -   @sap-ux/ui5-config@0.10.0
+    -   @sap-ux/odata-service-writer@0.9.4
+    -   @sap-ux/ui5-application-writer@0.9.4
+
 ## 0.9.3
 
 ### Patch Changes